wip: the big short
