+++ src/context/TransactionsProvider/TransactionsProvider.tsx
@@ -1,11 +1,5 @@
 import type { AccountId } from '@shapeshiftoss/caip'
-import {
-  cosmosChainId,
-  ethChainId,
-  foxAssetId,
-  fromAccountId,
-  osmosisChainId,
-} from '@shapeshiftoss/caip'
+import { ethChainId, foxAssetId, fromAccountId } from '@shapeshiftoss/caip'
 import type { Transaction } from '@shapeshiftoss/chain-adapters'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
 import { DefiProvider, DefiType } from 'features/defi/contexts/DefiManagerProvider/DefiCommon'
@@ -33,7 +27,6 @@ import {
   selectStakingOpportunitiesById,
 } from 'state/slices/selectors'
 import { txHistory } from 'state/slices/txHistorySlice/txHistorySlice'
-import { validatorDataApi } from 'state/slices/validatorDataSlice/validatorDataSlice'
 import { useAppDispatch } from 'state/store'
 
 import { usePlugins } from '../PluginProvider/PluginProvider'
@@ -162,12 +155,11 @@ export const TransactionsProvider: React.FC<TransactionsProviderProps> = ({ chil
             { wallet, accountType, accountNumber },
             msg => {
               const { getAccount } = portfolioApi.endpoints
-              const { getValidatorData } = validatorDataApi.endpoints
               const { onMessage } = txHistory.actions
 
-              // refetch validator data on new txs in case TVL or APR has changed
-              if ([cosmosChainId, osmosisChainId].includes(msg.chainId))
-                dispatch(getValidatorData.initiate(accountId))
+              // TODO(gomes): refetch Cosmsos opportunities in maybeRefetchOpportunities
+              // if ([cosmosChainId, osmosisChainId].includes(msg.chainId))
+              // dispatch(getValidatorData.initiate(accountId))
 
               // refetch account on new tx
               dispatch(
