+++ src/state/selectors.ts
@@ -7,8 +7,7 @@ import createCachedSelector from 're-reselect'
 import type { FiatRampAction } from 'components/Modals/FiatRamps/FiatRampsCommon'
 
 import type { ReduxState } from './reducer'
-import type { LpId, StakingId, UserStakingId } from './slices/opportunitiesSlice/types'
-import type { PubKey } from './slices/validatorDataSlice/validatorDataSlice'
+import type { LpId, StakingId, UserStakingId, ValidatorId } from './slices/opportunitiesSlice/types'
 
 /**
  * List of all the params filter consumed with selectParamFromFilter
@@ -24,7 +23,7 @@ type ParamFilter = Partial<{
   contractAddress: string
   assetId: AssetId
   accountId: AccountId
-  validatorAddress: PubKey // TODO(0xdef1cafe): change this to an AccountId that represents the chainId:validatorAddress
+  validatorId: ValidatorId
   accountNumber: number
   chainId: ChainId
   userStakingId: UserStakingId
@@ -53,7 +52,7 @@ export const selectAccountIdParamFromFilter = selectParamFromFilter('accountId')
 export const selectAccountNumberParamFromFilter = selectParamFromFilter('accountNumber')
 export const selectAssetIdParamFromFilter = selectParamFromFilter('assetId')
 export const selectChainIdParamFromFilter = selectParamFromFilter('chainId')
-export const selectValidatorAddressParamFromFilter = selectParamFromFilter('validatorAddress')
+export const selectValidatorIdParamFromFilter = selectParamFromFilter('validatorId')
 export const selectUserStakingIdParamFromFilter = selectParamFromFilter('userStakingId')
 export const selectStakingIdParamFromFilter = selectParamFromFilter('stakingId')
 export const selectLpIdParamFromFilter = selectParamFromFilter('lpId')
