+++ src/state/slices/selectors.ts
@@ -12,7 +12,6 @@ export * from './marketDataSlice/selectors'
 export * from './portfolioSlice/selectors'
 export * from './preferencesSlice/selectors'
 export * from './txHistorySlice/selectors'
-export * from './validatorDataSlice/selectors'
 export * from './opportunitiesSlice/selectors'
 
 /**
