+++ src/state/store.ts
@@ -18,7 +18,6 @@ import { opportunitiesApi } from './slices/opportunitiesSlice/opportunitiesSlice
 import { portfolioApi } from './slices/portfolioSlice/portfolioSlice'
 import * as selectors from './slices/selectors'
 import { txHistoryApi } from './slices/txHistorySlice/txHistorySlice'
-import { validatorDataApi } from './slices/validatorDataSlice/validatorDataSlice'
 
 const persistConfig = {
   key: 'root',
@@ -34,7 +33,6 @@ const apiMiddleware = [
   marketApi.middleware,
   assetApi.middleware,
   txHistoryApi.middleware,
-  validatorDataApi.middleware,
   foxyApi.middleware,
   swapperApi.middleware,
   fiatRampApi.middleware,
@@ -48,7 +46,6 @@ export const clearState = () => {
   store.dispatch(slices.assets.actions.clear())
   store.dispatch(slices.marketData.actions.clear())
   store.dispatch(slices.txHistory.actions.clear())
-  store.dispatch(slices.validatorData.actions.clear())
   store.dispatch(slices.portfolio.actions.clear())
   store.dispatch(slices.opportunities.actions.clear())
 
@@ -56,7 +53,6 @@ export const clearState = () => {
   store.dispatch(apiSlices.marketApi.util.resetApiState())
   store.dispatch(apiSlices.portfolioApi.util.resetApiState())
   store.dispatch(apiSlices.txHistoryApi.util.resetApiState())
-  store.dispatch(apiSlices.validatorDataApi.util.resetApiState())
   store.dispatch(apiSlices.opportunitiesApi.util.resetApiState())
 }
